+++ packages/core/src/config/index.ts
@@ -229,6 +229,10 @@ async function readConfigFile(
             filename: fileName,
             path: outputPath,
           },
+          external: [
+            ...module.builtinModules,
+            ...module.builtinModules.map((m) => `node:${m}`),
+          ],
           partialBundling: {
             moduleBuckets: [
               {
@@ -245,7 +249,7 @@ async function readConfigFile(
       });
       const compiler = new Compiler(normalizedConfig);
       await compiler.compile();
-      await compiler.writeResourcesToDisk();
+      compiler.writeResourcesToDisk();
 
       const filePath = path.join(outputPath, fileName);
       // Change to vm.module of node or loaders as far as it is stable
@@ -255,8 +259,12 @@ async function readConfigFile(
         return (await import(filePath)).default;
       }
     } else {
-      const config = (await import(resolvedPath)).default;
-      return config;
+      // Change to vm.module of node or loaders as far as it is stable
+      if (process.platform === 'win32') {
+        return (await import(pathToFileURL(resolvedPath).toString())).default;
+      } else {
+        return (await import(resolvedPath)).default;
+      }
     }
   }
 }